Break down tests to match new commands
